Select Region. For select and pick component.
The user defines a region from the stitched image. The section is cropped, and the center of the coordinate within the image is saved as a coordinate in a text file. This code will be run for 2 stitched images (one for picking and one for placing). The location will be saved and the stage will be moved to that location. The coordinate will need to be converted into physical units.